Various changes in GromacsTopologyParser.py
1) made it possible to incluue the itps in the same directory
as the top and gro files are located int. Before, it only looked
at the directory the program was run in.

2) there appear to be issues with the way atomstypes are read into atoms.
See below the type of change made:
             if len(split) == 3:
                               # Searching for matching bondtype to pull values from
-                            tempType = AbstractBondType(split[0], split[1], split[2])
+                            # Can't be this hard?
+                            atomtype1 = currentMolecule._atoms[int(split[0])-1].getAtomType()[0]
+                            atomtype2 = currentMolecule._atoms[int(split[1])-1].getAtomType()[0]
+                            tempType = AbstractBondType(atomtype1, atomtype2, split[2])
                             bondType = self.bondtypes.get(tempType)

So somehow the string text needs to be extracted from the atom type, and that should be
used to match the bond type.

Also, changes like this, explaining where bondType info comes fomr.

if isinstance(bondType, CubicBondType):
-                                split.append(length)
-                                split.append(C2)
-                                split.append(C3)
+                                split.append(bondType.length)
+                                split.append(bondType.C2)
+                                split.append(bondType.C3)

Finally,  there's apparently a dihedral format that just gives lists
the middle atoms, with wildcards assumed on the edge; this currently
isn't supported in the code; this will need to be fixed eventually.
Does the current code support wildcard matching in dihedrals at all?