1. Gromacs topology parser now accepts nonwildcard types of dihedraltypes. This can be checked with hostguest or 2PPN examples. However, bond types are reading in incorrectly.
2. added a new test case, micelledrug, which is what was in micelle/complex.top

3. Modified 2PPN.top so that it used the new 4.6 style includes.

4. Changed GromacsToplogyParser.py so that it can read in included files
of the form $GMXPATH/forcefieldname/includefile.itp.  Note there is a slight
danger in the current approach, in that if there are includefiles.itp that are
included in multiple directories that are used previously, it could choose a random one.
I couldn't get around this since an included file does not currently keep track of which
file it was incldued from, so it could inherit the parent directory from that.

5. Started playing around with regression test, then realized it would be better for Alex to handle that.